+++ dodola/services.py
@@ -66,7 +66,7 @@ def train_qdm(historical, reference, out, variable, kind):
 
 
 @log_service
-def apply_qdm(simulation, qdm, year, variable, out):
+def apply_qdm(simulation, qdm, year, variable, out, include_quantiles=False):
     """Apply trained QDM to adjust a year within a simulation, dump to NetCDF.
 
     Dumping to NetCDF is a feature likely to change in the near future.
@@ -86,6 +86,9 @@ def apply_qdm(simulation, qdm, year, variable, out):
     out : str
         fsspec-compatible path or URL pointing to NetCDF4 file where the
         QDM-adjusted simulation data will be written.
+    include_quantiles : bool
+        Flag to indicate whether bias-corrected quantiles should be
+        included in the QDM-adjusted output.
     """
     sim_ds = storage.read(simulation)
     qdm_ds = storage.read(qdm)
@@ -94,7 +97,11 @@ def apply_qdm(simulation, qdm, year, variable, out):
     variable = str(variable)
 
     adjusted_ds = adjust_quantiledeltamapping_year(
-        simulation=sim_ds, qdm=qdm_ds, year=year, variable=variable
+        simulation=sim_ds,
+        qdm=qdm_ds,
+        year=year,
+        variable=variable,
+        include_quantiles=include_quantiles,
     )
 
     # Write to NetCDF, usually on local disk, pooling and "fanning-in" NetCDFs is
@@ -327,7 +334,7 @@ def rechunk(x, target_chunks, out):
 
 
 @log_service
-def regrid(x, out, method, domain_file, weights_path=None):
+def regrid(x, out, method, domain_file, weights_path=None, astype=None):
     """Regrid climate data
 
     Parameters
@@ -342,6 +349,8 @@ def regrid(x, out, method, domain_file, weights_path=None):
         Storage URL to input xr.Dataset domain file to regrid to.
     weights_path : optional
         Local file path name to write regridding weights file to.
+    astype : str, numpy.dtype, or None, optional
+        Typecode or data-type to which the regridded output is cast.
     """
     ds = storage.read(x)
 
@@ -352,6 +361,7 @@ def regrid(x, out, method, domain_file, weights_path=None):
         ds_domain,
         method=method,
         weights_path=weights_path,
+        astype=astype,
     )
 
     storage.write(out, regridded_ds)