+++ dodola/tests/test_core.py
@@ -115,6 +115,36 @@ def test_adjust_quantiledeltamapping_year_kind(variable_kind, expected):
     assert all(adjusted_ds[target_variable] == expected)
 
 
+def test_adjust_quantiledeltamapping_include_quantiles():
+    """Test that include-quantiles flag results in bias corrected quantiles
+    included in output"""
+    target_variable = "fakevariable"
+    n_simdays = 5 * 365  # 100 years of daily simulation.
+
+    model_bias = 2.0
+    ts_sim = np.ones(n_simdays, dtype=np.float64)
+    sim = _timeseriesfactory(
+        ts_sim * model_bias, start_dt="2015-01-01", variable_name=target_variable
+    )
+
+    target_year = 2017
+
+    # Yes, I'm intentionally training the QDM to a different bias. This is to
+    # spurn a difference between "kind" adjustments...
+    qdm = _train_simple_qdm(
+        target_variable="fakevariable", kind="+", additive_bias=model_bias + 1
+    )
+    adjusted_ds = adjust_quantiledeltamapping_year(
+        simulation=sim,
+        qdm=qdm,
+        year=target_year,
+        variable=target_variable,
+        include_quantiles=True,
+    )
+    # check that quantiles are contained in output
+    assert "sim_q" in adjusted_ds[target_variable].coords
+
+
 def test_adjust_quantiledeltamapping_year_output_time():
     """Check 'time' year and edges of QDM adjusted output