+++ HISTORY.rst
@@ -5,6 +5,9 @@ History
 
 0.X.X (XXXX-XX-XX)
 ------------------
+* Add ``include-quantiles`` flag to ``apply_qdm`` to allow for including quantile information in bias corrected output. (PR #95, @dgergel)
+* Add precipitation unit conversion to ``standardize_gcm``. (PR #94, @dgergel)
+* Add ``astype`` argument to ``regrid``. (PR #92, @brews)
 * Make ``dodola`` container's default CMD. (PR #90, @brews)
 * Add ``dask-kubernetes``, ``distributed`` to container environment. (PR #90, @brews)
 * Improve subprocess and death handling in Docker container. (PR #90, @brews)